Prototype for displaying flight paths (#63)
* Clicking on airport marker shows airport name

* Implemented clearMap() function

* Prototype for displaying flight data

* Prototype for displaying flight paths
